2nd/main/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
pandas
sqlalchemy
psycopg2-binary
Jinja2
